test: cleanup tests, moved incentive liquidity test from Staking.test.ts to Incentive.test.ts
