+++ the-graph/subgraphs/staking/tests/Staking.test.ts
@@ -1,40 +1,18 @@
-import { Address, BigInt, log } from '@graphprotocol/graph-ts'
-import { test, assert, clearStore } from 'matchstick-as/assembly/index'
-import { onIncentiveCreated, onStake, onSubscribe, onUnstake, onUnsubscribe } from '../src/mappings/staking'
+import { Address, BigInt } from '@graphprotocol/graph-ts'
+import { assert, clearStore, test } from 'matchstick-as/assembly/index'
 import { getStakeId } from '../src/functions/staking'
-import { createIncentiveCreatedEvent, createStakeEvent, createSubscribeEvent, createUnstakeEvent, createUnsubscribeEvent } from './mocks'
+import { onStake, onUnstake } from '../src/mappings/staking'
+import { createStakeEvent, createUnstakeEvent } from './mocks'
 
 const ALICE = Address.fromString('0x00000000000000000000000000000000000a71ce')
-const BOB = Address.fromString('0x0000000000000000000000000000000000000b0b')
 const TOKEN = Address.fromString('0x0000000000000000000000000000000000000001')
-const INCENTIVE_ID = BigInt.fromString('1')
-const INITIAL_AMOUNT = BigInt.fromString('1000000')
-
-function setup(): void {
-  const REWARD_TOKEN = Address.fromString('0x0000000000000000000000000000000000000002')
-  let startTime = BigInt.fromString('1646068510')
-  let endTime = BigInt.fromString('1646075000')
-
-  let incentiveCreatedEvent = createIncentiveCreatedEvent(
-    TOKEN,
-    REWARD_TOKEN,
-    ALICE,
-    INCENTIVE_ID,
-    INITIAL_AMOUNT,
-    startTime,
-    endTime
-  )
-  onIncentiveCreated(incentiveCreatedEvent)
-}
 
 function cleanup(): void {
   clearStore()
 }
 
 test('Stake', () => {
-  setup()
-
-  const amount = BigInt.fromString("10000000")
+  const amount = BigInt.fromString('10000000')
   let stakeEvent = createStakeEvent(TOKEN, ALICE, amount)
 
   // When: Alice stakes an amount
@@ -50,17 +28,13 @@ test('Stake', () => {
   onStake(stakeEvent)
 
   // Then: The liquidity is doubled
-  assert.fieldEquals('Stake', stakeId, 'liquidity', amount.times(BigInt.fromString("2")).toString())
+  assert.fieldEquals('Stake', stakeId, 'liquidity', amount.times(BigInt.fromString('2')).toString())
 
   cleanup()
 })
 
-
-
 test('Stake and unstake', () => {
-  setup()
-
-  const amount = BigInt.fromString("10000000")
+  const amount = BigInt.fromString('10000000')
   let stakeEvent = createStakeEvent(TOKEN, ALICE, amount)
   let unstakeEvent = createUnstakeEvent(TOKEN, ALICE, amount)
 
@@ -78,30 +52,6 @@ test('Stake and unstake', () => {
 
   // Then: Stake entity liquidity is updated
   assert.fieldEquals('Stake', stakeId, 'liquidity', '0')
-
-  cleanup()
-})
-
-
-test('subscribe/unsubscribe updates the incentives staked liquidity', () => {
-  setup()
-  const amount = BigInt.fromString("10000000")
-  let stakeEvent = createStakeEvent(TOKEN, ALICE, amount)
-  let subscribeEvent = createSubscribeEvent(INCENTIVE_ID, ALICE)
-  let unsubscribeEvent = createUnsubscribeEvent(INCENTIVE_ID, ALICE)
-
-  assert.fieldEquals('Incentive', INCENTIVE_ID.toString(), 'liquidityStaked', '0')
-
-  onStake(stakeEvent)
-  // When: subscribe event is triggered
-  onSubscribe(subscribeEvent)
-
-  // Then: the incentives staked liquidity is increased
-  assert.fieldEquals('Incentive', INCENTIVE_ID.toString(), 'liquidityStaked', amount.toString())
-
-  // And: an unsubscribe decreases liquidity
-  onUnsubscribe(unsubscribeEvent)
-  assert.fieldEquals('Incentive', INCENTIVE_ID.toString(), 'liquidityStaked', '0')
 
   cleanup()
 })
